Review comments resolved
