--- packages/client/src/pages/leaderboard/consts.ts
@@ -1,7 +1,0 @@
-export const RATING_FIELD_NAME = 'leaderboardPoints';
-
-export const leaderboardRequest = {
-  ratingFieldName: RATING_FIELD_NAME,
-  cursor: 0,
-  limit: 10,
-};
